chore: create vertex.rs
