+++ crates/vide/src/api/mesh.rs
@@ -1,38 +1,9 @@
 use std::sync::MutexGuard;
 
-use bytemuck::{Pod, Zeroable};
 use wgpu::util::DeviceExt;
 
-use super::shader::Shader;
+use super::{shader::Shader, vertex::Vertex};
 use crate::render::Renderer;
-
-#[derive(Default, Debug, Clone, Copy, Pod, Zeroable)]
-#[repr(C)]
-pub struct Vertex {
-  pub position: [f32; 2],
-  pub uv: [f32; 2],
-}
-
-impl Vertex {
-  pub fn desc<'a>() -> wgpu::VertexBufferLayout<'a> {
-    wgpu::VertexBufferLayout {
-      array_stride: std::mem::size_of::<Vertex>() as wgpu::BufferAddress,
-      step_mode: wgpu::VertexStepMode::Vertex,
-      attributes: &[
-        wgpu::VertexAttribute {
-          offset: 0,
-          shader_location: 0,
-          format: wgpu::VertexFormat::Float32x2,
-        },
-        wgpu::VertexAttribute {
-          offset: std::mem::size_of::<[f32; 2]>() as wgpu::BufferAddress,
-          shader_location: 1,
-          format: wgpu::VertexFormat::Float32x2,
-        },
-      ],
-    }
-  }
-}
 
 #[derive(Debug)]
 pub struct Mesh {
